+++ components/eye_ui/eye_ui.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+// Call initArduino() in app_main() before calling this.
+void eye_ui_start(void);
+
+#ifdef __cplusplus
+}
+#endif